// src/store/modules/root/enum.ts
export enum EGetters {
  isEnabledSite = 'isEnabledSite',
  getUserData = 'getUserData',
}

export enum EMutations {
  initUiSettings = 'initUiSettings',
  initSiteSettings = 'initSiteSettings',
  initUserData = 'initUserData',
  toggleSiderCollapsed = 'toggleSiderCollapsed',
  setConcurrencyRequests = 'setConcurrencyRequests',
  setExpectTorrents = 'setExpectTorrents',
  setLocale = 'setLocale',
  toggleEnabledSite = 'toggleEnabledSite',
  updateUserData = 'updateUserData',
  updateSearchConfigs = 'updateSearchConfigs',
  deleteSearchConfigs = 'deleteSearchConfigs',
  addSearchConfigs = 'addSearchConfigs',
  setSelectedConfig = 'setSelectedConfig',
  setSearchText = 'setSearchText',
  setRunSearch = 'setRunSearch',
  setTimeout = 'setTimeout',
}

export enum EActions {
  initUiSettings = 'initUiSettings',
  initSiteSettings = 'initSiteSettings',
  initUserData = 'initUserData',
  toggleSiderCollapsed = 'toggleSiderCollapsed',
  setConcurrencyRequests = 'setConcurrencyRequests',
  setExpectTorrents = 'setExpectTorrents',
  setLocale = 'setLocale',
  toggleEnabledSite = 'toggleEnabledSite',
  updateUserData = 'updateUserData',
  updateSearchConfigs = 'updateSearchConfigs',
  deleteSearchConfigs = 'deleteSearchConfigs',
  addSearchConfigs = 'addSearchConfigs',
  setSelectedConfig = 'setSelectedConfig',
  setTimeout = 'setTimeout',
}